relocate PlatformHelperVersioning test
